crawler/requirements.txt
seleniumbase>=4.29.0
pdfplumber>=0.11.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.0.0
